feat(app): weather 컴포넌트 분리
