+++ apps/safers-app/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { useWeather } from "./useWeather";
+export { useWeather, type HourlyTemp } from "./useWeather";
