+++ apps/safers-app/src/pages/home/components/RightPanel.tsx
@@ -1,22 +1,20 @@
 import { useWeather } from "../../../hooks/useWeather";
-import {
-  extractWeatherState,
-  getWeatherIcon,
-  extractEnvironmentData,
-} from "../../../utils/weather";
+import { Spinner } from "@pf-dev/ui";
+import { Weather } from "./Weather";
 
 export function RightPanel() {
   const nx = Number(import.meta.env.VITE_SITE_NX);
   const ny = Number(import.meta.env.VITE_SITE_NY);
 
-  const { currentTemp, hourlyTemps, data } = useWeather({ nx, ny });
+  const { currentTemp, hourlyTemps, data, isLoading } = useWeather({ nx, ny });
 
-  const weatherState = data ? extractWeatherState(data) : { pty: null, sky: null };
-  const currentHour = new Date().getHours();
-  const weatherIcon = getWeatherIcon(weatherState.pty, weatherState.sky, currentHour);
-  const ncstData = data
-    ? extractEnvironmentData(data)
-    : { humidity: null, windDirection: "--", windSpeed: null };
+  if (isLoading) {
+    return (
+      <aside className="z-10 flex h-full w-[50rem] flex-shrink-0 items-center justify-center">
+        <Spinner size="lg" />
+      </aside>
+    );
+  }
 
   return (
     <aside className="z-10 flex h-full w-[50rem] flex-shrink-0 flex-col gap-4 p-4">
@@ -40,47 +38,7 @@ export function RightPanel() {
 
       {/* 현장 정보 + 날씨 */}
       <div className="rounded-lg border border-white/80 p-4 h-[12rem] shadow-[0_4px_10px_0_rgba(0,0,0,0.08)]">
-        <div className="flex gap-6 h-full">
-          <div className="bg-white rounded-xl w-38 h-full shrink-0 flex items-center justify-center">
-            <img
-              src={`${import.meta.env.VITE_CONTEXT_PATH}/assets/icons/${weatherIcon}`}
-              alt="weather"
-              className="w-20 h-20"
-            />
-          </div>
-          <div className="flex flex-col flex-1 justify-between">
-            <div className="flex gap-10 items-end">
-              <strong className="text-5xl">{currentTemp ?? "--"}°</strong>
-              <div className="flex items-center gap-2">
-                습도 {ncstData.humidity ?? "--"}%
-                {ncstData.windSpeed ? (
-                  <>
-                    <span className="w-1 h-1 bg-[#999999] rounded-full"></span>
-                    {ncstData.windDirection} {ncstData.windSpeed}m/s
-                  </>
-                ) : null}
-              </div>
-            </div>
-            <ul className="flex gap-3">
-              {hourlyTemps.map((item) => (
-                <li
-                  key={item.hour}
-                  className={`flex flex-col items-center min-w-13 p-2 gap-1 rounded-4xl shadow-[0_4px_10px_0_rgba(0,0,0,0.08)] text-sm ${
-                    item.isCurrent ? "bg-white/80 text-[#333333]" : "bg-white/50 text-[#9499B1]"
-                  }`}
-                >
-                  <span>{item.isCurrent ? "now" : `${item.hour}:00`}</span>
-                  <img
-                    src={`${import.meta.env.VITE_CONTEXT_PATH}/assets/icons/${item.isCurrent ? weatherIcon : getWeatherIcon(item.pty, item.sky, parseInt(item.hour, 10))}`}
-                    alt="weather"
-                    className="w-6 h-6"
-                  />
-                  <span className="font-bold">{item.temp}°</span>
-                </li>
-              ))}
-            </ul>
-          </div>
-        </div>
+        <Weather currentTemp={currentTemp} hourlyTemps={hourlyTemps} data={data} />
       </div>
 
       {/* 환경 데이터 */}
